# GoProgram/week4/web/docs/swagger.yaml
basePath: /
definitions:
  gin.H:
    additionalProperties: {}
    type: object
  models.ChangePasswordRequest:
    properties:
      new_password:
        minLength: 6
        type: string
      old_password:
        type: string
    required:
    - new_password
    - old_password
    type: object
  models.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  models.UpdateUserRequest:
    properties:
      nickname:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      nickname:
        type: string
      password:
        minLength: 6
        type: string
      updated_at:
        type: string
      username:
        maxLength: 20
        minLength: 3
        type: string
    required:
    - nickname
    - password
    - username
    type: object
  models.UserResponse:
    properties:
      created_at:
        type: string
      nickname:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 基于 Gin 框架的用户管理系统
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: 用户管理系统 API
  version: "1.0"
paths:
  /auth/check:
    get:
      description: 检查当前用户认证状态
      produces:
      - application/json
      responses:
        "200":
          description: 认证状态
          schema:
            $ref: '#/definitions/gin.H'
      summary: 检查认证状态
      tags:
      - users
  /login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 登录凭证
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 登录成功
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/gin.H'
      summary: 用户登录
      tags:
      - users
  /logout:
    post:
      description: 用户退出登录
      produces:
      - application/json
      responses:
        "200":
          description: 退出成功
          schema:
            $ref: '#/definitions/gin.H'
      summary: 用户退出
      tags:
      - users
  /register:
    post:
      consumes:
      - application/json
      description: 注册新用户
      parameters:
      - description: 用户信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: 注册成功
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/gin.H'
        "409":
          description: 用户已存在
          schema:
            $ref: '#/definitions/gin.H'
      summary: 用户注册
      tags:
      - users
  /user/info:
    get:
      description: 获取当前登录用户信息
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/models.UserResponse'
      summary: 获取用户信息
      tags:
      - users
    put:
      consumes:
      - application/json
      description: 更新当前用户昵称
      parameters:
      - description: 用户信息
        in: body
        name: userInfo
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的用户信息
          schema:
            $ref: '#/definitions/models.UserResponse'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/gin.H'
      summary: 更新用户信息
      tags:
      - users
  /user/password:
    put:
      consumes:
      - application/json
      description: 修改当前用户密码
      parameters:
      - description: 密码信息
        in: body
        name: password
        required: true
        schema:
          $ref: '#/definitions/models.ChangePasswordRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 密码修改成功
          schema:
            $ref: '#/definitions/gin.H'
        "400":
          description: 请求错误
          schema:
            $ref: '#/definitions/gin.H'
        "401":
          description: 原密码错误
          schema:
            $ref: '#/definitions/gin.H'
      summary: 修改密码
      tags:
      - users
  /users:
    get:
      description: 获取所有用户列表（需要登录）
      produces:
      - application/json
      responses:
        "200":
          description: 用户列表
          schema:
            items:
              $ref: '#/definitions/models.UserResponse'
            type: array
      summary: 获取所有用户
      tags:
      - users
swagger: "2.0"